+++ Assets/Scripts/Character/CookieGearController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 807421bf0fb4c7e4ea95e2d051c2e4aa
