+++ Assets/Scripts/Character/CookieHealthController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0ff69fd05f23e764b890da765b0baa50
