+++ Assets/Scripts/Character/CookieInputController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 445836af562c90745898acfbf1f1a570
