+++ Assets/Scripts/Character/CookieSizeController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 290b9fbd9cf74a548b2c9abb7a9490e9
